Fix reference to undefined `this.surface` property
Before making this change, my app doesn't render, and I see the following error in the Chrome Dev console:
```
Uncaught TypeError: Cannot call method 'add' of undefined - ScrollContainer.js:37
```
After making this change, the error disappears, and the ScrollContainer appears to render itself and behave as expected.

I might be ignorant of some step that occurs in `this.setOptions(options)` where `this.surface` might be defined. However, if you don't supply an `options` argument, the constructor always fails.

If there is a process for adding tests to submitted pull requests before they can be merged, I'd be happy to oblige.